Move HttpCredentialsHelper to main component/http package and give it package visibility.
