+++ components/camel-http/src/main/java/org/apache/camel/component/http/ProxyHttpClientConfigurer.java
@@ -16,7 +16,6 @@
  */
 package org.apache.camel.component.http;
 
-import org.apache.camel.component.http.helper.HttpCredentialsHelper;
 import org.apache.http.HttpHost;
 import org.apache.http.auth.Credentials;
 import org.apache.http.auth.NTCredentials;
