+++ components/camel-http/src/test/java/org/apache/camel/component/http/HttpCredentialsHelperTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.http.helper;
+package org.apache.camel.component.http;
 
 import org.apache.camel.test.junit5.CamelTestSupport;
 import org.apache.http.auth.AuthScope;
